@@ -7,14 +7,14 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "blockly": "^5.20210325.1",
-    "core-js": "^3.6.5",
-    "register-service-worker": "^1.7.1",
-    "sweetalert2": "^11.0.16",
-    "vue": "^3.0.0",
-    "vue-router": "^4.0.0-0",
-    "vuex": "^4.0.0-0"
+    "axios": "^1.6.2",
+    "blockly": "^10.2.2",
+    "core-js": "^3.33.2",
+    "register-service-worker": "^1.7.2",
+    "sweetalert2": "^11.10.0",
+    "vue": "^3.3.8",
+    "vue-router": "^4.2.5",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
